backend/requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
sqlalchemy==2.0.0
python-dotenv==1.0.0
google-generativeai==0.3.0
anthropic==0.7.0
apscheduler==3.10.0